requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.0
scikit-image>=0.18.0
scipy>=1.7.0
kornia>=0.6.0
kornia-moons>=0.2.0
lightglue>=0.1.0 